Correcting the defaults in cache schema
