+++ internal/controller/olsconfig_controller.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 
 	"github.com/go-logr/logr"
-	olsv1alpha1 "github.com/openshift/lightspeed-operator/api/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -30,6 +29,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
+
+	olsv1alpha1 "github.com/openshift/lightspeed-operator/api/v1alpha1"
 )
 
 // OLSConfigReconciler reconciles a OLSConfig object
@@ -82,6 +83,11 @@ func (r *OLSConfigReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 	err = r.reconcileAppServer(ctx, olsconfig)
 	if err != nil {
 		r.logger.Error(err, "Failed to reconcile application server")
+		return ctrl.Result{}, err
+	}
+	err = r.reconcileRedisServer(ctx, olsconfig)
+	if err != nil {
+		r.logger.Error(err, "Failed to reconcile ols redis")
 		return ctrl.Result{}, err
 	}
 	r.logger.Info("reconciliation done", "olsconfig generation", olsconfig.Generation)
