+++ internal/controller/types.go
@@ -3,8 +3,9 @@ package controller
 import (
 	"context"
 
-	olsv1alpha1 "github.com/openshift/lightspeed-operator/api/v1alpha1"
 	"k8s.io/apimachinery/pkg/util/intstr"
+
+	olsv1alpha1 "github.com/openshift/lightspeed-operator/api/v1alpha1"
 )
 
 /*** controller inernal ***/
@@ -60,7 +61,7 @@ type LoggingConfig struct {
 }
 
 type ConversationCacheConfig struct {
-	// Type of cache to use. Default: "memory"
+	// Type of cache to use. Default: "redis"
 	Type string `json:"type" default:"redis"`
 	// Redis cache configuration
 	Redis RedisCacheConfig `json:"redis,omitempty"`
